# perf: Optimize airspace pipeline - 40% faster Continental AU loading